+++ gemstone/core/handlers.py
@@ -93,7 +93,7 @@ class TornadoJsonRpcHandler(RequestHandler):
           was already sent to the client.
 
         :param single_request: A :py:class:`dict` object representing a Request object
-        :return: A :py:class:`dict` object representing a Response object or None
+        :return: A :py:class:`dict` object representing a Response object or None if no response is expected
 
         """
         is_notification = False
@@ -236,6 +236,13 @@ class TornadoJsonRpcHandler(RequestHandler):
         self.write_response(error=err)
 
     def prepare_method_call(self, method, args):
+        """
+        Wraps a method so that method() will call method(*args) or method(**args), depending of args type
+
+        :param method: a callable object (method)
+        :param args: dict or list with the parameters for the function
+        :return: a 'patched' callable
+        """
         if isinstance(args, list):
             to_call = partial(method, *args)
         elif isinstance(args, dict):